configure background export types in dialog
So far, the type of background export determined extra export types.
Originally this increased the number of export types by a factor of 2,
now (with the addition of "no ruling") by a factor of 3. This is a lot
already and does not scale.

Instead, reduce the export types to those determined by the file type and
configure the background export types in the export dialog, along with
progressive mode and quality parameters (depending on file type).

In addition to a cleaner UI this makes it easier to remember the last
choice or make the default configurable (to be implemented by a future
series) rather than forcing the user to make the same choice over and
over again.