fix position highlight clipping
The position hightlight is a filled circle of radius up to 30 pixels,
possibly with a border stroke of width up to 30 pixels. Therefore, it
extends up to 90 pixels in width and height.

Currently, the code draws the circle on a 60x60 canvas, possibly
clipping the border if 2*radius + linewidth > 60. Change this to a
90x90 canvas which fits all cases.

We could compute a tight canvas, but the excess area is transparent
anyway, and memory usage is not an issue here.