specify gtest dependency in spec but turn gtest off.
269167da ("Migrate from CppUnit to GoogleTest (#3476)", 2021-11-08)
switched from cppunit to gtest and removed the cppunit dependency. It
did not put in any gtest dependency because that commit makes the build
download the testsuite from github and install it along with the build
executable, which is probably the most stupd way to do it.

So, turn it off for now.

A proper fix would use the system gtest, or include gtest src and
prevent installation (https://github.com/google/googletest/issues/868).