implement selection paste anchor
Implement an anchor which allows quick on-screen pasting of selections.
While this does not act on the current selection (but pastes the
clipboard) it is still useful: conceptually, paste belongs to
cut/copy/duplicate, so the anchor sits where it is expected.
practically, it allows to cut or copy "somewhere" and paste this close
to another selection (possibly quickly selected by tap if enabled).

The anchors are drawn so that duplicate = copy+paste both visually as
well as in terms of coordinate vectors.