fix crash with zero width stroke segments
553588a4c ("Split away helper functions from view/StrokeView",
2022-07-21) split away helper functions but also changed behaviour of
the code. In particular, it introduced an assert for non-zero width of
stroke segments of strokes with pressure which was not there before. As
a consequence, xournalpp crashes on files on which it did not cross
before.

Zero width stroke segments can happen for strokes with pressure due to
fading as well as rounding. So, remove the assert which was introduced
by the split.