Do not build the stacktrace wrapper
01734d85f ("Add a wrapper around the executable that informs of crashes (#6563)", 2025-08-07)
introduced a stacktrace wrapper which requires external sources, being
pulled in during the build. This is a no-go on Fedora.

We do want to keep build with debug info, so rather than switching to a
Release build, disable the wrapper. We may want to switch to
RelWithDebInfo in the future (or release if necessary).